DOMA-3719 New version of notification
