+++ apps/condo/domains/common/hooks/useHotCodeReload.tsx
@@ -1,0 +1,44 @@
+import React, { useCallback, useEffect } from 'react'
+import Router from 'next/router'
+import { notification } from 'antd'
+import { useIntl } from '@core/next/intl'
+import { Button } from '@condo/domains/common/components/Button'
+import { useCurrentBuild } from './useCurrentBuild'
+import { usePrevious } from './usePrevious'
+
+
+
+export function useHotCodeReload (): void {
+    const intl = useIntl()
+    const NotificationTitle = intl.formatMessage({ id: 'HotCodeReload.title' })
+    const NotificationMessage = intl.formatMessage({ id: 'HotCodeReload.message' })
+    const ButtonLabel = intl.formatMessage({ id: 'HotCodeReload.action' })
+
+    const buildId = useCurrentBuild()
+    const previousBuildId = usePrevious<string>(buildId)
+
+    const handleNotificationClose = useCallback(() => {
+        Router.reload()
+    }, [])
+
+    useEffect(() => {
+        if (buildId && previousBuildId && previousBuildId !== buildId) {
+            const btn: React.ReactNode = (
+                <Button
+                    type={'sberDefaultGradient'}
+                    onClick={handleNotificationClose}
+                >
+                    {ButtonLabel}
+                </Button>
+            )
+            notification.success({
+                key: buildId,
+                btn,
+                message: NotificationTitle,
+                description: NotificationMessage,
+                duration: null,
+                closeIcon: React.Fragment,
+            })
+        }
+    }, [buildId, previousBuildId, NotificationTitle, NotificationMessage, ButtonLabel, handleNotificationClose])
+}
