+++ apps/condo/lang/en/en.json
@@ -70,8 +70,9 @@
   "example.Email": "name@example.org",
   "example.Name": "Jon Johnson",
   "example.Phone": "+1 (300) 300 22 22",
-  "HotReload.title": "System update!",
-  "HotReload.message": "We have updated the system, so this page will now reload",
+  "HotCodeReload.title": "System update",
+  "HotCodeReload.message": "We've improved some features, but you'll need to reload the page for the changes to take effect",
+  "HotCodeReload.action": "Got it, reload",
   "ExportAsExcel": "Export to Excel",
   "FiltersLabel": "Filters",
   "field.AccountNumberShort": "Account Number",
