+++ apps/condo/lang/ru/ru.json
@@ -70,8 +70,9 @@
   "example.Email": "name@example.org",
   "example.Name": "Иванов Сергей",
   "example.Phone": "+7 (900) 800 22 22",
-  "HotReload.title": "Обновление системы!",
-  "HotReload.message": "Мы обновили систему, поэтому данная страница сейчас перезагрузится",
+  "HotCodeReload.title": "Обновление системы",
+  "HotCodeReload.message": "Мы улучшили некоторые функции, но чтобы изменения заработали — страницу нужно перезагрузить",
+  "HotCodeReload.action": "Понятно, перезагрузить",
   "ExportAsExcel": "Выгрузить в Excel",
   "FiltersLabel": "Фильтры",
   "field.AccountNumberShort": "ЛС",
